Move compiling prost functions and types into prost module
